Option to add intercept variable to linear models.
Intercept variable can be added to the models. public API was created for now until option setter is added.
Tests were also changed from expected results to computing the gradient of the model and checking it is close enough to 0.
A bug in the logistic regression objective is fixed for cases with multiple features.

Change-Id: If90d340c74431c7aee79dffcc88e895c99050066