Updating the CSV and option setting code so they work together; this mainly involved switching the CSV reader to use a class, of which the original parser struct is a member.
The main changes to the option setting code are that now non-categorical string options are allowed.

There's a bit of an ugly hack in da_options_set_int to allow the skiprow functionality to work (it's an integer option that is appended to the end of an array), but I'll leave it as it is for now, until I think of a better way.

debug_build

Change-Id: I08e93d48ec935c3ed39aca9c97e72d775b602fa8