Improvements to basic statistics routines
The main change here is to allow a choice of unbiased (n_obs-1) or biased (n_obs) estimators for standard deviations and variances.
This has been applied to:
- da_variance
- da_covariance_matrix
- da_standardize

It has also been applied to the PCA API as an extra option.
Related to this, I have switched to the biased estimator for the variances that are returned with the skewness and kurtosis routines.
This purely as it's a little more natural and fits better with the formulae used for skewness and kurtosis.

Finally, standardize API has been updated to allow 'unstandardization' and returning of mean and standard deviation arrays.

Tests and doc for all the above updated where appropriate.

BUILD_DOC
JIRA-ID - [CPUPL-3544]

Change-Id: I657940ab43f7a6b48c66e96299f980122c568e02
Reviewed-on: https://gerrit-git.amd.com/c/cpulibraries/er/aocl-da/+/961562
Tested-by: Osi Request <Osi.Request@amd.com>
Tested-by: Llvm Cpu <llvm.cpu@amd.com>
Reviewed-by: Edvin Hopkins <Edvin.Hopkins@amd.com>