Enable floating point arguments on Python API constructors and remove need to specifying precision
The sklearn APIs work by creating both double and single precision handles and discarding one when we know the precision.
The Python APIs don't do this and as a result can't have floating point values on the constructor (because the underlying pybind11 class constructors are not templated).
This patchset moves the sklearn trick into the Python API. This makes the Python APIs more intuitive and simpolfies the sklearn patches.
Also some fixes to get rid of compilation and pytest warnings.

Change-Id: Idb0e8966bcb9b2bd3b16a1959f20594237fe794c
Reviewed-on: http://gerrit-git.amd.com/c/cpulibraries/er/aocl-da/+/1139634
Tested-by: Osi Request <Osi.Request@amd.com>
Tested-by: Llvm Cpu <llvm.cpu@amd.com>
Reviewed-by: Edvin Hopkins <Edvin.Hopkins@amd.com>