Remove the basic_handle() constructor
But add an alternative constructor where error and options objects can be supplied by the calling code.
This means e.g. the random forest code can still remain as lightweight as possible, but a developer would have to explicitly pass in nullptrs to the constructor so things should be a bit safer.

Change-Id: I218cb52f50a71a9a85bc73599c4b59dc03b37871
Reviewed-on: http://gerrit-git.amd.com/c/cpulibraries/er/aocl-da/+/1142021
Tested-by: Osi Request <Osi.Request@amd.com>
Tested-by: Llvm Cpu <llvm.cpu@amd.com>
Reviewed-by: Benjamin Marteau <Benjamin.Marteau@amd.com>